+++ wagtailsocialfeed/wagtail_hooks.py
@@ -5,13 +5,13 @@ from django.core.urlresolvers import reverse
 from django.utils import six
 from django.utils.translation import ugettext_lazy as _
 
-from wagtail.contrib.modeladmin.menus import SubMenu
 from wagtail.contrib.modeladmin.options import ModelAdmin, modeladmin_register
-from wagtail.wagtailadmin.menu import MenuItem, SubmenuMenuItem
+from wagtail.wagtailadmin.menu import Menu, MenuItem, SubmenuMenuItem
 from wagtail.wagtailcore import hooks
 
 from . import urls
 from .models import SocialFeedConfiguration
+from django.db.models.signals import post_save
 
 
 class SocialFeedConfigurationAdmin(ModelAdmin):
@@ -22,6 +22,52 @@ class SocialFeedConfigurationAdmin(ModelAdmin):
     add_to_settings_menu = True
     list_display = ('source', 'username')
     list_filter = ('source', )
+
+
+class SocialFeedModerateMenu(Menu):
+    config_menu_items = {
+        # Contains all submenu items mapped to the id's of the
+        # `SocialFeedConfiguration` they link to
+    }
+
+    def __init__(self):
+        # Iterate over existing configurations that are moderated
+        config_qs = SocialFeedConfiguration.objects.filter(moderated=True)
+        for config in config_qs:
+            self.config_menu_items[config.id] = \
+                self._create_moderate_menu_item(config)
+
+        self._registered_menu_items = self.config_menu_items.values()
+        self.construct_hook_name = None
+
+        post_save.connect(self._update_menu, sender=SocialFeedConfiguration)
+
+    def _update_menu(self, instance, **kwargs):
+        """
+        Call whenever a `SocialFeedCongiration` gets changed.
+
+        When it is not moderated anymore, but exists in our menu, it should be
+        removed.
+        When it is moderated but does not exist yet, we should create a
+        new menu item.
+        """
+        has_menu_item = instance.id in self.config_menu_items.keys()
+
+        if not instance.moderated and has_menu_item:
+            menu_item = self.config_menu_items.pop(instance.id)
+            index = self._registered_menu_items.index(menu_item)
+            del self._registered_menu_items[index]
+        elif instance.moderated and not has_menu_item:
+            menu_item = self._create_moderate_menu_item(instance)
+            self.config_menu_items[instance.id] = menu_item
+            self._registered_menu_items.append(menu_item)
+
+    def _create_moderate_menu_item(self, config):
+        """Create a submenu item for the moderate admin page."""
+        url = reverse('wagtailsocialfeed:moderate', kwargs={'pk': config.pk})
+        return MenuItem(six.text_type(config), url,
+                        classnames='icon icon-folder-inverse')
+
 
 modeladmin_register(SocialFeedConfigurationAdmin)
 
@@ -36,16 +82,8 @@ def register_admin_urls():
 
 @hooks.register('register_admin_menu_item')
 def register_socialfeed_menu():
-    config_menu_items = []
-    config_qs = SocialFeedConfiguration.objects.filter(moderated=True)
-    for config in config_qs:
-        url = reverse('wagtailsocialfeed:moderate', kwargs={'pk': config.pk})
-        config_menu_items.append(
-            MenuItem(six.text_type(config), url,
-                     classnames='icon icon-folder-inverse')
-        )
-
-    socialfeed_menu = SubMenu(config_menu_items)
+    # Create the main socialfeed menu item
+    socialfeed_menu = SocialFeedModerateMenu()
     return SubmenuMenuItem(
         _('Social feed'), socialfeed_menu, classnames='icon icon-fa-rss',
         order=800)