+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 0.1.dev2 (xxxx.xx.xx)
 ---------------------
 + Added block type 'SocialFeedBlock'
++ Live refresh moderate menu items when configuration gets changed
 
 
 0.1.dev1 (2016.09.01)